requirements.txt
google-api-python-client~=2.100.0
httplib2~=0.22.0
oauth2client~=4.1.3
python-dotenv~=1.0.0
python-telegram-bot==13.7